Add behavior to merge kiali-bot PRs for version bumps
kiali-bot may place PRs to bump version numbers once when a release is created. However, the PR needs to be approved and merged manually. This adds a behavior so that the kiali-bot app approves and merges those PRs without the need of human intervention.

The JiraIssueExporter behavior is, for now, deactivated since it is only a PoC; not ready to be deployed.