Remove 'standard' dev dependency
Standard is a Javascript linter. It is in the initial project stucture provided when running create-probot-app. It has some support for typescript, but for some reason it was not working and, also, it is not customizable (unless using standardx). Replacing it with ESlint and prettier, since this seems more popular in the community.