PrChecker diagnosing query
Apparently, the list returned by getOrQueryPrsForCommit is sometimes empty.
Adding log messages to see if GitHub is returning an empty query or if conditionals are not OK.

Also, adding making the bot to not fail with a never ending "in progress" check. At least, the check should now finish with failure status, making it possible to request a re-check.